Fix documentation issues
