+++ packages/text-bitmap/src/BitmapFont.js
@@ -15,10 +15,6 @@ const available = {};
  */
 export class BitmapFont
 {
-    /**
-     * @param {PIXI.BitmapFontData} data
-     * @param {Object.<string, PIXI.Texture>|PIXI.Texture[]} textures
-     */
     constructor(data, textures)
     {
         const [info] = data.info;
@@ -157,56 +153,3 @@ export class BitmapFont
         return available;
     }
 }
-
-/**
- * Normalized parsed data from .fnt files.
- *
- * @class
- * @memberof PIXI
- */
-export class BitmapFontData
-{
-    constructor()
-    {
-        /**
-         * @member {Array.<{ face:string, size:number }>}
-         * @readOnly
-         */
-        this.info = [];
-
-        /**
-         * @member {Array.<{ lineHeight:number }>}
-         * @readOnly
-         */
-        this.common = [];
-
-        /**
-         * @member {Array.<{ id:number, file:string }>}
-         * @readOnly
-         */
-        this.page = [];
-
-        /**
-         * @member {Array.<{
-         *   id:string,
-         *   page:number,
-         *   x:number,
-         *   y:number,
-         *   width:number,
-         *   height:number,
-         *   xoffset:number,
-         *   yoffset:number,
-         *   xadvance:number
-         * }>}
-         * @readOnly
-         */
-        this.char = [];
-
-        /**
-         * @member {Array.<{first:number, second:number, amount:number}>}
-         * @readOnly
-         */
-        this.kerning = [];
-    }
-}
-
