+++ packages/text-bitmap/src/index.js
@@ -1,3 +1,4 @@
 export * from './BitmapText';
 export * from './BitmapFontLoader';
 export * from './BitmapFont';
+export * from './BitmapFontData';
